+++ NEWS.md
@@ -1,6 +1,8 @@
 # longurl 0.3.2
 
 * Ensure input is character before processing. (@JohnCoene, #4)
+* Removed dplyr & purrr deps
+* As a result of ^^ also removed progress bar 
 
 # longurl 0.3.1